on some coook shit ong trust we got that module on us type shit

+++ ingest/model/gr24_pedal.go
@@ -9,10 +9,10 @@ type GR24Pedal struct {
 	Millis    int       `json:"millis" gorm:"index"`
 	CreatedAt time.Time `json:"created_at" gorm:"autoCreateTime;precision:6"`
 	// Frame 1 - ID:200
-	APPSOne            int `json:"apps_one"`
-	APPSTwo            int `json:"apps_two"`
-	BrakePressureFront int `json:"brake_pressure_front"`
-	BrakePressureRear  int `json:"brake_pressure_rear"`
+	APPSOne            float64 `json:"apps_one"`
+	APPSTwo            float64 `json:"apps_two"`
+	BrakePressureFront float64 `json:"brake_pressure_front"`
+	BrakePressureRear  float64 `json:"brake_pressure_rear"`
 }
 
 func (GR24Pedal) TableName() string {
