+++ ingest/model/gr24_vdm.go
@@ -13,11 +13,11 @@ type GR24VDM struct {
 	DataType   string `json:"data_type"`
 	Periodic   string `json:"periodic"`
 	// Frame 2 - ID:101
-	FanOneSpeed   int `json:"fan_one_speed"`
-	FanTwoSpeed   int `json:"fan_two_speed"`
-	FanThreeSpeed int `json:"fan_three_speed"`
-	FanFourSpeed  int `json:"fan_four_speed"`
-	PumpSpeed     int `json:"pump_speed"`
+	FanOneSpeed   float64 `json:"fan_one_speed"`
+	FanTwoSpeed   float64 `json:"fan_two_speed"`
+	FanThreeSpeed float64 `json:"fan_three_speed"`
+	FanFourSpeed  float64 `json:"fan_four_speed"`
+	PumpSpeed     float64 `json:"pump_speed"`
 	// Frame 3 - ID:102
 	SoundCode string `json:"sound_code"`
 	// Frame 4 - ID:202
@@ -25,27 +25,27 @@ type GR24VDM struct {
 	// Frame 5 - ID:250
 	State string `json:"state" gorm:"index"`
 	// Frame 6 - ID:278
-	CurrentAC int `json:"current_ac"`
+	CurrentAC float64 `json:"current_ac"`
 	// Frame 7 - ID:534
-	TargetBrakeCurrent int `json:"target_brake_current"`
+	TargetBrakeCurrent float64 `json:"target_brake_current"`
 	// Frame 8 - ID:790
-	TargetERPM int `json:"target_erpm"`
+	TargetERPM float64 `json:"target_erpm"`
 	// Frame 9 - ID:1046
-	TargetPosition int `json:"target_position"`
+	TargetPosition float64 `json:"target_position"`
 	// Frame 10 - ID:1302
-	RelativeCurrentAC int `json:"relative_current_ac"`
+	RelativeCurrentAC float64 `json:"relative_current_ac"`
 	// Frame 11 - ID:1558
-	RelativeBrakeCurrentAC int `json:"relative_brake_current_ac"`
+	RelativeBrakeCurrentAC float64 `json:"relative_brake_current_ac"`
 	// Frame 12 - ID:1814
 	DigitalIO string `json:"digital_io"`
 	// Frame 13 - ID:2070
-	MaxCurrentAC int `json:"max_current_ac"`
+	MaxCurrentAC float64 `json:"max_current_ac"`
 	// Frame 14 - ID:2326
-	MaxBrakeCurrentAC int `json:"max_brake_current_ac"`
+	MaxBrakeCurrentAC float64 `json:"max_brake_current_ac"`
 	// Frame 15 - ID:2582
-	MaxCurrentDC int `json:"max_current_dc"`
+	MaxCurrentDC float64 `json:"max_current_dc"`
 	// Frame 16 - ID:2838
-	MaxBrakeCurrentDC int `json:"max_brake_current_dc"`
+	MaxBrakeCurrentDC float64 `json:"max_brake_current_dc"`
 	// Frame 17 - ID:3094
 	DriveEnable string `json:"drive_enable"`
 }
